Acceptance tests

+++ app/controllers/questions_controller.rb
@@ -1,27 +1,38 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: %i[index show]
+
   expose :questions, -> { Question.all }
   expose :question
+  expose :answer, -> { Answer.new }
 
   def create
-    @question = Question.create(question_params)
+    @question = current_user.questions.create(question_params)
     if @question.save
-      redirect_to @question
+      redirect_to @question, notice: 'Your question successfully created.'
     else
       render :new
     end
   end
 
   def update
-    if question.update(question_params)
-      redirect_to question
+    if current_user.author_of? question
+      if question.update(question_params)
+        redirect_to question
+      else
+        render :edit
+      end
     else
-      render :edit
+      render :edit, notice: "You're not an author of this question"
     end
   end
 
   def destroy
-    question.destroy
-    redirect_to questions_path
+    if current_user.author_of?(question)
+      question.destroy
+      redirect_to questions_path, notice: 'Your question successfully deleted.'
+    else
+      render :show
+    end
   end
 
   private
